backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Jamie Eriksson <jamieeriksson@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
python-dotenv = "^0.15.0"
django-cors-headers = "^3.7.0"
django-countries = "^7.0"
djangorestframework-simplejwt = "^4.6.0"
dj-database-url = "^0.5.0"
factory-boy = "^3.2.0"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.6"
Faker = "^6.5.2"
django-filter = "^2.4.0"
requests = "^2.25.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^5.0.0"}
pytest-django = "^4.1.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "==20.8b1"
pytest-cov = "^2.11.1"
pytest-django = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
